autodocs/requirements.txt
gh-md-to-html==1.21.2
pygments==2.12.0
mistune==2.0.0rc1
pyyaml==6.0
